+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-# 1.0.0 (2021-11-17)
+# 1.0.0 (2021-11-18)
 
 
 ### Features
 
-* add core code ([be67d45](https://github.com/hsycc/easy-jsbridge-brigde/commit/be67d452dc3c3043cfbeda01fdf4c15924ae1053))
-* add webpack to build browser version ([46994e8](https://github.com/hsycc/easy-jsbridge-brigde/commit/46994e8262852216cc0a08cb849773c23329d43f))
+* add core code ([be67d45](https://github.com/hsycc/poly-jsbridge-brigde/commit/be67d452dc3c3043cfbeda01fdf4c15924ae1053))
+* add webpack to build browser version ([46994e8](https://github.com/hsycc/poly-jsbridge-brigde/commit/46994e8262852216cc0a08cb849773c23329d43f))
+* build changes ([5e7ebc0](https://github.com/hsycc/poly-jsbridge-brigde/commit/5e7ebc0e59913071d9480e76c545d9b7237503b8))
+* rename package name , update  test cases, update doc ([7edace0](https://github.com/hsycc/poly-jsbridge-brigde/commit/7edace0df04a06fc89f208973d5c12ea85946255))